use ray for creating child process
