+++ python/pyproject.toml
@@ -10,7 +10,7 @@ classifiers = [
     "Programming Language :: Python :: Implementation :: CPython",
     "Programming Language :: Python :: Implementation :: PyPy",
 ]
-version = "2.0.0-beta12"
+version = "2.0.0-beta13"
 dependencies = [
     "openpyxl==3.1.2",
     "boto3==1.28.9",
@@ -21,7 +21,7 @@ dependencies = [
     "pandas==2.2.2",
     "pyarrow==16.1.0",
     "pystalk==0.7.0",
-    "ray==2.24.0",
+    "ray==2.31.0",
     "ruamel.yaml==0.17.31",
     "scikit-allel==1.3.8",
     "scikit-learn==1.2.2",
@@ -36,6 +36,7 @@ dependencies = [
     "somadata==1.0.0",
     "statsmodels==0.14.2",
     "nest-asyncio==1.6.0",
+    "grpcio==1.64.1",
 ]
 
 [tool.maturin]
